requirements.txt
Django>=3.0,<3.1
wagtail>=2.9,<2.10
django-allauth
git+git://github.com/LogicalAddress/wagtailstreamforms.git@master
mysqlclient
django-dotenv==1.4.2
wagtailmenus==3.0.2
wagtail-contact-reply
django-countries
wagtail-treemodeladmin
django-crispy-forms
crispy-forms-foundation==0.8.0
django-storages
boto3
djangorestframework
django-cors-headers
git+https://github.com/gbozee/django-paystack.git@master#egg=paystack
wagtailmedia
django-termsandconditions
wagalytics
wagtail-metadata